moved schemas to api directory
